Move Beam to privacy coins
As much as I would prefer Grin to be the one awesome new privacy coin, I
believe Beam has a good chance of actually getting there in a much much
faster time. They are producing results.

Among others, most dominant work in progress is the merger of [Lelantus
and Mimblewimble](https://github.com/BeamMW/beam/wiki/Lelantus-MW), i.e.
the new ZCoin protocol and Mimblewimble. While anonymity-set might be
not as high as ZCash, the relative simplicity of either protocol and the
already strong promises make this the goto technology for privacy in the
future, IMHO. Of course, I can only look into so many communities.